Prevent trivial .mat file changes
Previously, after running genuslab in the editor, there would often be
small changes to g2paintmaterial.mat and h2paintmaterial.mat, because
setting the paint coordinates in PaintUV() of PaintableTexture.cs
resulted in those coordinates being saved to disk.

Now we make a copy of each paint material in
ReplaceTextureWithRenderTexture() of PaintableTexture.cs and use that
copy for all paint operations.  Thus the changes are discarded when the
application exits, and never written to disk.