Move laser toggle functionality into StickBehavior
Instead of separate toggle script, maintain the laser state in
StickBehavior.

Now, "drawing" and "high beam" are collapsed to a single state called
"active".  It is controlled by StickBehavior.makeActive() and
StickBehavior.makeInactive().  Currently the control script ties this to
the right index trigger button.

Due to this change, the StickBehavior script needs to be able to find
the beam object.  To do this, the beam object should be tagged with the
new tag "Beam".