Fix missing texture on g2 object
Two corrections to accomplish this:

1) Set the _MainTex property of the material SSOITMaterial to the
default six-panel texture.  This "marks" it as a texture that needs to
be linked to the shared RenderTexture by PaintableTexture.

2) Modify Start() of PaintableTexture so that it knows to look for
materials that are not directly applied to the object, but are instead
stored as a .ssoitMaterial property of a ssoitControl script attached to
it.  This makes sure the proper linking still happens if
ssoitControl.Start() runs after PaintableTexture.Start().  (If they run
in the other order, there is no problem because the ssoitMaterial will
already be applied as the main texture when PaintableTexture does the
linking.)