Fix problem with line "jumping" from previous position
Drawing on g2 works, drawing on H2 still broken.

This commit fixes a problem where if you start drawing a second time,
the first thing that is drawn is a solid line from the previous position
to the current one.  The issue was that lastPosition needed to be reset
to (NaN,NaN,NaN) when the beam was set to inactive, but this was not
done.