Disable copying of paint materials in PaintableTexture
The copy feature was added to stop Unity from saving various changes
to disk.  However it introduced an unexpected race condition between
Start() methods of PaintableTexture and h2viewcontrol.

If h2viewcontrol.Start ran first, it would store the original paint
texture of the h2 object.  Later, PaintableTexture would duplicate that
and set the copy would be used for painting.  The result was that
h2viewcontrol attempting to set properties of the paint shader would
fail, resulting in painting that did not line up with the laser.